1. Bug fixes as on 05/07/2020 1:00pm

+++ app/src/main/java/dm/sime/com/kharetati/view/fragments/HomeFragment.java
@@ -374,7 +374,7 @@ public class HomeFragment extends Fragment implements GridMenuAdapter.OnMenuSele
         builder.append("remarks=" + Global.getPlatformRemark() + "&");
         String lang = Global.CURRENT_LOCALE.compareToIgnoreCase("en") == 0 ? "en" : "ar";
         builder.append("lng=" + lang + "&");
-        builder.append("fontSize=" + Global.fontSize * Global.fontScale+ "&");
+        builder.append("fontSize=" + (int)(Global.fontSize * Global.fontScale)+ "&");
         builder.append("appsrc=kharetati&");
         if(!Global.isUserLoggedIn){
             //Guest
